q1/requirements.txt
flask>=3.0
boto3>=1.28.0
pyjwt>=2.8.0
gunicorn>=21.2.0
